+++ src/finufft3d.cpp
@@ -7,8 +7,8 @@
 #include <iostream>
 #include <iomanip>
 
-int finufft3d1(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,int iflag,
-	       double eps, BIGINT ms, BIGINT mt, BIGINT mu, dcomplex* fk,
+int finufft3d1(INT nj,double* xj,double *yj,double *zj,dcomplex* cj,int iflag,
+	       double eps, INT ms, INT mt, INT mu, dcomplex* fk,
 	       nufft_opts opts)
  /*  Type-1 3D complex nonuniform FFT.
 
@@ -25,7 +25,7 @@ int finufft3d1(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,int iflag
      used, otherwise the - sign is used, in the exponential.
                            
    Inputs:
-     nj     number of sources (integer of type BIGINT; see utils.h)
+     nj     number of sources (integer of type INT; see utils.h)
      xj,yj,zj   x,y,z locations of sources on 3D domain [-pi,pi]^3.
      cj     size-nj complex double array of source strengths, 
             (ie, stored as 2*nj doubles interleaving Re, Im).
@@ -56,16 +56,16 @@ int finufft3d1(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,int iflag
   spread_opts spopts;
   int ier_set = setup_kernel(spopts,eps,opts.R);
   if (ier_set) return ier_set;
-  BIGINT nf1; set_nf_type12(ms,opts,spopts,&nf1);
-  BIGINT nf2; set_nf_type12(mt,opts,spopts,&nf2);
-  BIGINT nf3; set_nf_type12(mu,opts,spopts,&nf3);
+  INT64 nf1; set_nf_type12((BIGINT)ms,opts,spopts,&nf1);
+  INT64 nf2; set_nf_type12((BIGINT)mt,opts,spopts,&nf2);
+  INT64 nf3; set_nf_type12((BIGINT)mu,opts,spopts,&nf3);
   if (nf1*nf2*nf3>opts.maxnalloc) {
     fprintf(stderr,"nf1*nf2*nf3=%.3g exceeds maxnalloc of %.3g\n",(double)nf1*nf2*nf3,(double)opts.maxnalloc);
     return ERR_MAXNALLOC;
   }
   cout << scientific << setprecision(15);  // for debug
 
-  if (opts.debug) printf("3d1: (ms,mt,mu)=(%ld,%ld,%ld) (nf1,nf2,nf3)=(%ld,%ld,%ld) nj=%ld ...\n",ms,mt,mu,nf1,nf2,nf3,nj); 
+  if (opts.debug) printf("3d1: (ms,mt,mu)=(%ld,%ld,%ld) (nf1,nf2,nf3)=(%ld,%ld,%ld) nj=%ld ...\n",(INT64)ms,(INT64)mt,(INT64)mu,nf1,nf2,nf3,(INT64)nj); 
 
   // STEP 0: get DCT of half of spread kernel in each dim, since real symm:
   CNTime timer; timer.start();
@@ -115,8 +115,8 @@ int finufft3d1(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,int iflag
   return 0;
 }
 
-int finufft3d2(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,
-	       int iflag,double eps, BIGINT ms, BIGINT mt, BIGINT mu,
+int finufft3d2(INT nj,double* xj,double *yj,double *zj,dcomplex* cj,
+	       int iflag,double eps, INT ms, INT mt, INT mu,
 	       dcomplex* fk, nufft_opts opts)
 
  /*  Type-2 3D complex nonuniform FFT.
@@ -128,7 +128,7 @@ int finufft3d2(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,
                        -mu/2 <= k3 <= (mu-1)/2
 
    Inputs:
-     nj     number of sources (integer of type BIGINT; see utils.h)
+     nj     number of sources (integer of type INT; see utils.h)
      xj,yj,zj     x,y,z locations of sources on 3D domain [-pi,pi]^3.
      fk     double complex array of Fourier series values (size ms*mt*mu,
             increasing fastest in ms to slowest in mu, ie Fortran ordering).
@@ -158,16 +158,16 @@ int finufft3d2(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,
   spread_opts spopts;
   int ier_set = setup_kernel(spopts,eps,opts.R);
   if (ier_set) return ier_set;
-  BIGINT nf1; set_nf_type12(ms,opts,spopts,&nf1);
-  BIGINT nf2; set_nf_type12(mt,opts,spopts,&nf2);
-  BIGINT nf3; set_nf_type12(mu,opts,spopts,&nf3);
+  INT64 nf1; set_nf_type12((BIGINT)ms,opts,spopts,&nf1);
+  INT64 nf2; set_nf_type12((BIGINT)mt,opts,spopts,&nf2);
+  INT64 nf3; set_nf_type12((BIGINT)mu,opts,spopts,&nf3);
   if (nf1*nf2*nf3>opts.maxnalloc) {
     fprintf(stderr,"nf1*nf2*nf3=%.3g exceeds maxnalloc of %.3g\n",(double)nf1*nf2*nf3,(double)opts.maxnalloc);
     return ERR_MAXNALLOC;
   }
   cout << scientific << setprecision(15);  // for debug
 
-  if (opts.debug) printf("3d2: (ms,mt,mu)=(%ld,%ld,%ld) (nf1,nf2,nf3)=(%ld,%ld,%ld) nj=%ld ...\n",ms,mt,mu,nf1,nf2,nf3,nj);
+  if (opts.debug) printf("3d2: (ms,mt,mu)=(%ld,%ld,%ld) (nf1,nf2,nf3)=(%ld,%ld,%ld) nj=%ld ...\n",(INT64)ms,(INT64)mt,(INT64)mu,nf1,nf2,nf3,(INT64)nj);
 
   // STEP 0: get Fourier coeffs of spread kernel in each dim:
   CNTime timer; timer.start();
@@ -219,8 +219,8 @@ int finufft3d2(BIGINT nj,double* xj,double *yj,double *zj,dcomplex* cj,
   return 0;
 }
 
-int finufft3d3(BIGINT nj,double* xj,double* yj,double *zj, dcomplex* cj,
-	       int iflag, double eps, BIGINT nk, double* s, double *t,
+int finufft3d3(INT nj,double* xj,double* yj,double *zj, dcomplex* cj,
+	       int iflag, double eps, INT nk, double* s, double *t,
 	       double *u, dcomplex* fk, nufft_opts opts)
  /*  Type-3 3D complex nonuniform FFT.
 
@@ -229,7 +229,7 @@ int finufft3d3(BIGINT nj,double* xj,double* yj,double *zj, dcomplex* cj,
                j=0
                           for k=0,...,nk-1
    Inputs:
-     nj     number of sources (integer of type BIGINT; see utils.h)
+     nj     number of sources (integer of type INT; see utils.h)
      xj,yj,zj   x,y,z location of sources in R^3.
      cj     size-nj complex double array of source strengths
             (ie, interleaving Re & Im parts)
@@ -267,24 +267,25 @@ int finufft3d3(BIGINT nj,double* xj,double* yj,double *zj, dcomplex* cj,
   spread_opts spopts;
   int ier_set = setup_kernel(spopts,eps,opts.R);
   if (ier_set) return ier_set;
-  BIGINT nf1,nf2,nf3;
+  INT64 nf1,nf2,nf3;
   double X1,C1,S1,D1,h1,gam1,X2,C2,S2,D2,h2,gam2,X3,C3,S3,D3,h3,gam3;
   cout << scientific << setprecision(15);  // for debug
 
   // pick x, s intervals & shifts, then apply these to xj, cj (twist iii)...
   CNTime timer; timer.start();
-  arraywidcen(nj,xj,&X1,&C1);  // get half-width, center, containing {x_j}
-  arraywidcen(nk,s,&S1,&D1);   // {s_k}
-  arraywidcen(nj,yj,&X2,&C2);  // {y_j}
-  arraywidcen(nk,t,&S2,&D2);   // {t_k}
-  arraywidcen(nj,zj,&X3,&C3);  // {z_j}
-  arraywidcen(nk,u,&S3,&D3);   // {u_k}
+  arraywidcen((BIGINT)nj,xj,&X1,&C1);  // get half-width, center, containing {x_j}
+  arraywidcen((BIGINT)nk,s,&S1,&D1);   // {s_k}
+  arraywidcen((BIGINT)nj,yj,&X2,&C2);  // {y_j}
+  arraywidcen((BIGINT)nk,t,&S2,&D2);   // {t_k}
+  arraywidcen((BIGINT)nj,zj,&X3,&C3);  // {z_j}
+  arraywidcen((BIGINT)nk,u,&S3,&D3);   // {u_k}
   // todo: if C1<X1/10 etc then set C1=0.0 and skip the slow-ish rephasing?
   set_nhg_type3(S1,X1,opts,spopts,&nf1,&h1,&gam1);          // applies twist i)
   set_nhg_type3(S2,X2,opts,spopts,&nf2,&h2,&gam2);
   set_nhg_type3(S3,X3,opts,spopts,&nf3,&h3,&gam3);
-  if (opts.debug) printf("3d3: X1=%.3g C1=%.3g S1=%.3g D1=%.3g gam1=%g nf1=%ld X2=%.3g C2=%.3g S2=%.3g D2=%.3g gam2=%g nf2=%ld X3=%.3g C3=%.3g S3=%.3g D3=%.3g gam3=%g nf3=%ld nj=%ld nk=%ld...\n",
-	 X1,C1,S1,D1,gam1,nf1,X2,C2,S2,D2,gam2,nf2,X3,C3,S3,D3,gam3,nf3,nj,nk);
+  if (opts.debug)
+    printf("3d3: X1=%.3g C1=%.3g S1=%.3g D1=%.3g gam1=%g nf1=%ld X2=%.3g C2=%.3g S2=%.3g D2=%.3g gam2=%g nf2=%ld X3=%.3g C3=%.3g S3=%.3g D3=%.3g gam3=%g nf3=%ld nj=%ld nk=%ld...\n",
+	   X1,C1,S1,D1,gam1,nf1,X2,C2,S2,D2,gam2,nf2,X3,C3,S3,D3,gam3,nf3,(INT64)nj,(INT64)nk);
   if (nf1*nf2*nf3>opts.maxnalloc) {
     fprintf(stderr,"nf1*nf2*nf3=%.3g exceeds maxnalloc of %.3g\n",(double)nf1*nf2*nf3,(double)opts.maxnalloc);
     return ERR_MAXNALLOC;
@@ -325,7 +326,7 @@ int finufft3d3(BIGINT nj,double* xj,double* yj,double *zj, dcomplex* cj,
     tp[k] = h2*gam2*(t[k]-D2);                         // so that |t'_k| < pi/R
     up[k] = h3*gam3*(u[k]-D3);                         // so that |u'_k| < pi/R
   }
-  int ier_t2 = finufft3d2(nk,sp,tp,up,fk,iflag,eps,nf1,nf2,nf3,fw,opts);
+  int ier_t2 = finufft3d2(nk,sp,tp,up,fk,iflag,eps,(INT)nf1,(INT)nf2,(INT)nf3,fw,opts);
   free(fw);
   if (opts.debug) printf("total type-2 (ier=%d):\t %.3g s\n",ier_t2,timer.elapsedsec());
   if (ier_t2>0) exit(ier_t2);
